Module rustorm::dao

Structs

DaoResult
meta result of a query useful when doing complex query, and also with paging TODO: good name: DaoRows

Enums

Type
supported generic datatypes for an ORM

Value
supported generic datatypes for an ORM

Traits

FromValue

IsDao
trait for converting dao to model sized and clonable

ToCompact
Ignore Column are columns that are redundant when displaying as API results

ToValue
rename to ToValue

Type Definitions

Dao
TODO: optimization, used enum types for the key values This will save allocation of string to enum keys which is a few bytes, int

ParseError