Trait rustorm::dao::IsDao

pub trait IsDao {
    fn from_dao(dao: &Dao) -> Self;
    fn to_dao(&self) -> Dao;
}

trait for converting dao to model sized and clonable

Required Methods

fn from_dao(dao: &Dao) -> Self

convert dao to an instance of the corresponding struct of the model taking into considerating the renamed columns

fn to_dao(&self) -> Dao

convert from an instance of the struct to a dao representation to be saved into the database