Module rustorm::database

Enums

BuildMode
specifies if the sql will be build in debug mode for debugging purposed

DbError

SqlOption
SqlOption, contains the info about the features and quirks of underlying database

Traits

Database
Generic Database interface This is the database interface which will should be implemented to you the specifics of each database platform At least all methods on this trait should be implemented for target deployment database A lower level API for manipulating objects in the database

DatabaseDDL
Deployment Database should implement this trait, to enable automated installation of the app, regardless what database platform the app is developed from.

DatabaseDev
implement this for database that you use as your development platform, to extract meta data information about the tables and their relationship to each other