Enum rustorm::database::SqlOption

pub enum SqlOption {
    UsesNumberedParam,
    UsesQuestionMark,
    SupportsReturningClause,
    SupportsCTE,
    SupportsInheritance,
    UsesSchema,
    ReturnMetaColumns,
}

SqlOption, contains the info about the features and quirks of underlying database

Variants

UsesNumberedParam

use the numbered parameters, as the case with rust-postgres

UsesQuestionMark

sqlite, jdbc

SupportsReturningClause

postgresql supports returning clause on insert and update

SupportsCTE

support CTE (common table expression ie. WITH) (postgresql, sqlite)

SupportsInheritance

supports inheritance (postgresql)

UsesSchema

whether the database uses schema (postgresl, oracle)

ReturnMetaColumns

wheter the returned rows in a query included Meta columns for easy extraction of records (postgres returns this), sqlite does not return meta columns, so you have to extract it by index yourself.