Struct rustorm::database::Pool

pub struct Pool {
    // some fields omitted
}

This pool contains database that are not necessarily same platform and configs database pools are stored using treemap, with the key is the url of the database owns the database, and only lend out reference when api tries to connect to the database TODO: protect the free form here with locks

Methods

impl Pool

fn init() -> Self

initialize a pool for database connection container call only once per application

fn reserve(url: &str, n: usize) -> Self

fn reserve_connection(&mut self, url: &str, n: usize) -> &mut Self

reserve a number of connections using the url config

fn get_db_with_url(&mut self, url: &str) -> Result<Platform, String>

exposed api to get connection from a pooled connection

fn release(&mut self, platform: Platform) -> &mut Self

release the used connection back to the free pool

fn total_connections(&self) -> usize

fn total_free_connections(&self) -> usize

return the number of free available connection in the pool

fn total_used_connections(&self) -> usize

return the number of free available connection in the pool