Struct rustorm::platform::postgres::Postgres

pub struct Postgres {
    pub pool: Option<PooledConnection<PostgresConnectionManager>>,
}

Fields

pool

a connection pool is provided

Methods

impl Postgres

Build the Query into a SQL statements that is a valid PostgreSQL sql query, TODO: support version SqlOptions/specific syntax

fn new() -> Self

create an instance, but without a connection yet, useful when just building sql queries specific to this platform inexpensive operation, so can have multiple instances

fn with_pooled_connection(pool: PooledConnection<PostgresConnectionManager>) -> Self

fn get_connection(&self) -> &Connection