Struct rustorm::query::builder::QueryBuilder

pub struct QueryBuilder {
    // some fields omitted
}

Methods

impl QueryBuilder

fn WITH(&mut self, query: Query, alias: &str) -> &mut Self

if the database support CTE declareted query i.e WITH, then this query will be declared if database doesn't support WITH queries, then this query will be wrapped in the from_query build a builder for this

fn SELECT() -> Self

fn SELECT_ALL() -> Self

fn INSERT() -> Self

fn UPDATE(to_table_name: &ToTableName) -> Self

fn DELETE() -> Self

fn DISTINCT(&mut self) -> &mut Self

add DISTINCT ie: SELECT DISTINCT

fn ALL(&mut self) -> &mut Self

fn SET(&mut self, column: &str, to_value: &ToValue) -> &mut Self

fn COLUMNS(&mut self, to_operand: &[&ToOperand]) -> &mut Self

fn VALUES(&mut self, to_values: &[&ToValue]) -> &mut Self

fn FROM(&mut self, to_source_field: &ToSourceField) -> &mut Self

A more terse way to write the query only 1 table is supported yet

fn INTO(&mut self, table: &ToTableName) -> &mut Self

into is used in rust, os settled with into_

fn TABLE(&mut self, table: &ToTableName) -> &mut Self

can be used in behalf of into_, from,

fn JOIN(&mut self, join: Join) -> &mut Self

join a table on this query

fn LEFT_JOIN(&mut self, join: Join) -> &mut Self

join a table on this query

fn RIGHT_JOIN(&mut self, join: Join) -> &mut Self

fn FULL_JOIN(&mut self, join: Join) -> &mut Self

fn INNER_JOIN(&mut self, join: Join) -> &mut Self

fn WHERE(&mut self, filter: Filter) -> &mut Self

fn GROUP_BY(&mut self, to_operand: &ToOperand) -> &mut Self

fn HAVING(&mut self, filter: Filter) -> &mut Self

fn ORDER_BY(&mut self, to_order: &ToOrder) -> &mut Self

fn LIMIT(&mut self, n: usize) -> &mut Self

fn OFFSET(&mut self, o: usize) -> &mut Self

fn build(&mut self, db: &Database) -> SqlFrag

build the query only, not executed, useful when debugging

fn collect_one<T: IsDao + IsTable>(&mut self, db: &Database) -> Result<T, DbError>

fn collect<T: IsDao + IsTable>(&mut self, db: &Database) -> Result<Vec<T>, DbError>