Struct rustorm::query::Query

pub struct Query {
    pub sql_type: SqlType,
    pub distinct: bool,
    pub enumerate_all: bool,
    pub declared_query: BTreeMap<String, Query>,
    pub enumerated_fields: Vec<Field>,
    pub distinct_on_columns: Vec<String>,
    pub from: Vec<SourceField>,
    pub joins: Vec<Join>,
    pub filters: Vec<Filter>,
    pub order_by: Vec<Order>,
    pub group_by: Vec<Operand>,
    pub having: Vec<Filter>,
    pub excluded_columns: Vec<ColumnName>,
    pub range: Range,
    pub values: Vec<Operand>,
    pub enumerated_returns: Vec<Field>,
}

Fields

sql_type

sql type determine which type of query to form, some fields are not applicable to other types of query

distinct

whether to select the records distinct

enumerate_all

whether to enumate all columns in involved models

declared_query

enumerated_fields

fields can be functions, column sql query, and even columns

distinct_on_columns

specify to use distinct ON set of columns

from

where the focus of values of column selection this is the table to insert to, update to delete, create, drop whe used in select, this is the pub from_table:Option, from field, where field can be a query, table, or function

joins

joining multiple tables

filters

filter records, ~ where statement of the query

order_by

ordering of the records via the columns specified

group_by

grouping columns to create an aggregate

having

having field,

excluded_columns

exclude the mention of the columns in the SQL query, useful when ignoring changes in update/insert records

range

caters both limit->offset and page->page_size setting page and page_size can not interchange the order

values

The data values, used in bulk inserting, updating,

enumerated_returns

the returning clause of the query when supported,

Methods

impl Query

fn new() -> Self

the default query is select

fn select() -> Self

fn insert() -> Self

fn update() -> Self

fn delete() -> Self

fn enumerate_all(&mut self)

fn all(&mut self)

fn select_all() -> Self

fn column(&mut self, column: &str)

all enumerated columns shall be called from this any conflict of columns from some other table will be automatically renamed columns that are not conflicts from some other table, but is the other conflicting column is not explicityly enumerated will not be renamed

fn columns(&mut self, columns: Vec<&str>)

fn exclude_column(&mut self, column: &str)

exclude columns when inserting/updating data also ignores the column when selecting records useful for manipulating thin records by excluding huge binary blobs such as images

fn exclude_columns(&mut self, columns: Vec<&str>)

fn distinct_on_columns(&mut self, columns: &Vec<String>)

fn value(&mut self, value: &ToValue)

fn set(&mut self, column: &str, value: &ToValue)

set a value of a column when inserting/updating records

fn set_limit(&mut self, limit: usize)

fn set_offset(&mut self, offset: usize)

fn get_range(&self) -> Range

fn only_from(&mut self, table: &ToTableName)

enumerate only the columns that is coming from this table this will invalidate enumerate_all

fn from_query(&mut self, query: Query, alias: &str)

a query to query from use WITH (query) t1 SELECT from t1 declaration in postgresql, sqlite use SELECT FROM (query) in oracle, mysql, others alias of the table

fn from(&mut self, to_source_field: &ToSourceField)

fn table(&mut self, to_source_field: &ToSourceField)

fn get_from_table(&self) -> Option<TableName>

returns the first table in the from clause

fn get_involved_tables(&self) -> Vec<TableName>

fn finalize(&mut self) -> &Self

preprocess the missing fields of the query, such as mentioning the columns of the from_table enumerate the columns of the involved tables skipping those which are explicitly ignored the query will then be built and ready to be executed if no enumerated fields and no excluded columns do a select all

fn enumerate_from_table(&mut self, table: &TableName)

fn get_renamed_columns(&self) -> Vec<(ColumnName, String)>

return the list of renamed columns, used in dao conversion to struc types

fn get_enumerated_columns(&self) -> Vec<&ColumnName>

return the list of enumerated columns will be used for updating records

fn add_filter(&mut self, filter: Filter)

fn add_filters(&mut self, filters: &Vec<Filter>)

fn filter_eq(&mut self, column: &str, value: &ToValue)

column = value

fn filter_lt(&mut self, column: &str, value: &ToValue)

column < value

fn filter_lte(&mut self, column: &str, value: &ToValue)

column <= value

fn filter_gt(&mut self, column: &str, value: &ToValue)

column > value

fn filter_gte(&mut self, column: &str, value: &ToValue)

column <= value

fn return_all(&mut self)

fn returns(&mut self, columns: Vec<&str>)

fn enumerate_column_as_return(&mut self, column: &str)

fn build(&mut self, db: &Database) -> SqlFrag

build the query only, not executed, useful when debugging

fn debug_build(&mut self, db: &Database) -> SqlFrag

Warning: don't use this in production

fn retrieve(&mut self, db: &Database) -> Result<DaoResult, DbError>

retrieve a generic types, type is unknown at runtime expects a return, such as select, insert/update with returning clause

fn retrieve_one(&mut self, db: &Database) -> Result<Option<Dao>, DbError>

expects a return, such as select, insert/update with returning clause no casting of data to structs is done This is used when retrieving multiple models in 1 query, then casting the records to its equivalent structs

fn execute(&mut self, db: &Database) -> Result<usize, DbError>

delete, update without caring for the return

fn collect<T: IsDao + IsTable>(&mut self, db: &Database) -> Result<Vec<T>, DbError>

execute the query, then convert the result

fn collect_one<T: IsDao + IsTable>(&mut self, db: &Database) -> Result<T, DbError>

execute the query then collect only 1 record