Struct rustorm::table::Column

pub struct Column {
    pub table: Option<String>,
    pub name: String,
    pub data_type: Type,
    pub db_data_type: String,
    pub is_primary: bool,
    pub is_unique: bool,
    pub default: Option<Operand>,
    pub comment: Option<String>,
    pub not_null: bool,
    pub foreign: Option<Foreign>,
    pub is_inherited: bool,
}

Fields

table

name

data_type

the generic data type, ie: u32, f64, string

db_data_type

the database data type of this column, ie: int, numeric, character varying

is_primary

is_unique

default

comment

not_null

foreign

is_inherited

determines if the column is inherited from the parent table

Methods

impl Column

fn nullable(&self) -> bool

fn complete_name(&self) -> String

fn corrected_name(&self) -> String

some column names may be a rust reserve keyword, so have to correct them

fn displayname(&self) -> String

fn condense_name(&self) -> String

shorten, compress the name based on the table it points to parent_organization_id becomes parent