Struct rustorm::dao::Dao

pub struct Dao {
    pub values: HashMap<String, Type>,
}

Fields

values

Methods

impl Dao

fn new() -> Self

fn set(&mut self, column: &str, value: &ToType)

fn set_value(&mut self, column: &str, value: Type)

fn get_value(&self, column: &str) -> Type

fn remove<T>(&mut self, column: &str) -> T where T: FromType

take the value and remove the content

fn get<T>(&self, column: &str) -> T where T: FromType

take the value but not removing the content

fn get_opt<T>(&self, column: &str) -> Option<T> where T: FromType

get optional value

fn get_ref(&self, column: &str) -> &Type

get a reference of the type

fn resolve_renamed_columns(&mut self, in_table: &str, renamed_columns: &BTreeMap<String, Vec<(String, String)>>)

set the short column names using the renamed columns from the table specified will be used when casting a generic dao to multiple dao values, ie. useful when querying 1 time using 1:1 joins TODO: deal with the optional columns