Struct rustorm::em::EntityManager

pub struct EntityManager<'a> {
    pub db: &'a Database,
}

A higher level API for manipulating objects in the database

Fields

db

Methods

impl<'a> EntityManager<'a>

fn new(db: &'a Database) -> Self

Create an entity manager with the database connection provided

fn create_table(&self, table: &Table)

create a database table aligned to this table definition

fn create_schema(&self, schema: &str)

create a schema or namespace in the database

fn delete(&self, table: &Table, filters: &Vec<Filter>) -> usize

delete records of this table

fn drop_table(&self, table: &Table)

drop the database table

fn drop_schema(&self, schema: &str)

drop the database schema

fn truncate_table(&self, table: &Table) -> usize

empty the database table

fn exist_table(&self, table: &Table) -> bool

determine if the table exist

fn exist_schema(&self, schema: &str) -> bool

determine if the schema exist

fn get_all<T>(&self) -> Vec<T> where T: IsTable + IsDao

get all the records of this table

fn get_all_only_columns<T>(&self, columns: Vec<&str>) -> Vec<T> where T: IsTable + IsDao

get all the records of this table, but return only the columns mentioned

fn get_all_ignore_columns<T>(&self, ignore_columns: Vec<&str>) -> Vec<T> where T: IsTable + IsDao

get all the records of this table, ignoring the columns listed, mentioned the other else

fn get_all_distinct(&self, table: &Table) -> Vec<Dao>

get all the distinct records of this table

fn get_all_with_filter<T>(&self, table: &Table, filters: Vec<Filter>) -> Vec<T> where T: IsTable + IsDao

get all the records on this table which passed thru the filters any query that specified more than the parameters should use the query api

fn get_one<T>(&self, filter: Filter) -> Vec<T> where T: IsTable + IsDao

get the first records of this table that passed thru the filters

fn get_exact<T>(&self, id: &ToType) -> T where T: IsTable + IsDao

get an exact match, the value is filter against the primary key of the table

Examples

extern crate rustorm;
extern crate uuid;
extern crate chrono;
extern crate rustc_serialize;
 
use rustorm::db::postgres::Postgres;
use uuid::Uuid;
 
use rustorm::em::EntityManager;
use gen::bazaar::Product;
 
mod gen;
  
 
fn main(){
    let pg= Postgres::with_connection("postgres://postgres:p0stgr3s@localhost/bazaar_v6");
    let em = EntityManager::new(&pg);
    let pid = Uuid::parse_str("6db712e6-cc50-4c3a-8269-451c98ace5ad").unwrap();
    let prod: Product = em.get_exact(&pid);
    println!("{}  {}  {:?}", prod.product_id, prod.name.unwrap(), prod.description);
}

fn insert<T>(&self, dao: Dao) -> T where T: IsTable + IsDao

insert this records to the database, return the inserted dao with values from default columns included

Example

extern crate bazaar;
extern crate rustorm;
 
use rustorm::em::EntityManager;
use rustorm::db::Postgres;
use rustorm::dao::Dao;
use bazaar::gen::bazaar::Product;
fn main(){
let pg = Postgres::with_connection("postgres://postgres:p0stgr3s@localhost/bazaar_v6");
match pg{
    Ok(pg) => {
        let em = EntityManager::new(&pg);
        let mut dao = Dao::new();
        dao.set("name", &"inserting 1 records");
        dao.set("description", &"testing insert 1 record to product");
        let dao = em.insert(&Product::table(), dao);
        let prod = Product::from_dao(&dao);
        println!("created: {}", prod.created);
    }
    Err(error) =>{
        println!("{}",error);
    }
}
}

fn insert_with_ignore_columns<T>(&self, dao: Dao, ignore_columns: Vec<&str>) -> T where T: IsTable + IsDao

insert this record on the database, ignoring some columns which are set by the database default columns that are ignored are set by the database automatically

fn insert_ignore_defaulted_columns<T>(&self, dao: Dao) -> T where T: IsTable + IsDao

insert this record on the database, explicitly setting the defaults of the columns it may produce the same result with insert_with_ignore_columns the query is different since it may mentions created now(),

fn reset(&self)

this is called when there is a problem with the transaction

fn rollback(&self)

when there is a problem with the transaction process, this can be called

fn update<T>(&self, dao: &Dao) -> T where T: IsTable + IsDao

update the Dao, return the updated Dao

fn update_ignore_columns<T>(&self, dao: &Dao, ignore_columns: Vec<&str>) -> T where T: IsTable + IsDao

update the Dao, return the updated Dao ignored columns will remain unchanged

fn update_only_columns<T>(&self, dao: &Dao, columns: Vec<&str>) -> T where T: IsTable + IsDao

update the Dao, return the updated Dao only the columns specified, the rest is unchanged

fn update_ignore_defaulted_columns<T>(&self, dao: &Dao) -> T where T: IsTable + IsDao

update the Dao, return the updated Dao the default columns will be reset to whatever the db's default function will come up. ie. updated column will be defaulted everytime a record is updated.

fn update_with_filter<T>(&self, dao: &Dao, filter: Vec<Filter>) -> T where T: IsTable + IsDao

update the Dao with filter, return the updated Dao

fn save<T>(&self, dao: T) -> T where T: IsTable + IsDao

whether to use insert or update insert when it is a new record update when it is an existing recor may use UPSERT in newer versions of postgres may use MERGE in oracle, mssql