Struct rustorm::table::Table

pub struct Table {
    pub schema: String,
    pub name: String,
    pub parent_table: Option<String>,
    pub sub_table: Vec<String>,
    pub comment: Option<String>,
    pub columns: Vec<Column>,
}

Fields

schema

which schema this belongs

name

the table name

parent_table

the parent table of this table when inheriting (>= postgresql 9.3) [FIXME] need to tell which schema this parent table belongs there might be same table in different schemas

sub_table

what are the other table that inherits this [FIXME] need to tell which schema this parent table belongs there might be same table in different schemas

comment

comment of this table

columns

columns of this table

Methods

impl Table

fn complete_name(&self) -> String

return the long name of the table using schema.table_name

fn struct_name(&self) -> String

capitalize the first later, if there is underscore remove it then capitalize the next letter

fn displayname(&self) -> String

get the display name of this table product_availability -> Product Availability

fn condensed_displayname(&self, table: &Table) -> String

get a shorter display name of a certain table when being refered to this table example product.product_availability -> Availability user.user_info -> Info

fn condensed_member_name(&self, used_in_table: &Table) -> String

get a condensed name of this table when used in contex with another table

fn has_column_name(&self, column: &str) -> bool

determine if this table has a colum named

fn primary_columns(&self) -> Vec<&Column>

return all the primary columns of this table

fn uninherited_columns(&self) -> Vec<&Column>

return all the columns of this table excluding the inherited columns

fn inherited_columns(&self) -> Vec<&Column>

return all the inherited columns

fn is_primary(&self, column_name: &str) -> bool

check to see if the column is a primary or not the Column.is_primary property is not reliable since it also list down the foreign key which makes it 2 entries in the table

fn unique_columns(&self) -> Vec<&Column>

return all the unique keys of this table

fn foreign_columns(&self) -> Vec<&Column>

fn get_table<'a>(schema: &str, table_name: &str, tables: &'a Vec<Table>) -> &'a Table

return the first match of table name regardless of which schema it belongs to. get the table definition using the table name from an array of table object [FIXME] Needs to have a more elegant solution by using HashMap

fn get_all_referenced_table<'a>(&'a self, all_tables: &'a Vec<Table>) -> Vec<RefTable>

fn to_tabledef_source_code(&self) -> (Vec<String>, String)

build a source code which express it self as a table object which is a meta definition of the struct itself